business/requirements.txt
yt-dlp
ffmpeg-python
faster-whisper
chromadb
sentence-transformers
ollama
fastapi
uvicorn
Pillow
tqdm